Fixed bug in month checking code of getData function
Found a bug in my fancy new month checker in the getData function, lower case months did not match the sentence case months in the array.

Changed the array to be lower case and function now converts the form input to lower case before checking.

To keep it looking clean on the front end, I set the input of the month to capitalise, it looks a little funny when a word is spelt wrong but it functions well enough.